Changed Double to Decimal for PR #311

+++ modules/flowable-form-engine/src/main/java/org/flowable/form/engine/impl/cmd/GetVariablesFromFormSubmissionCmd.java
@@ -114,7 +114,7 @@ public class GetVariablesFromFormSubmissionCmd implements Command<Map<String, Ob
                 result = null;
             }
         }
-            else if (formField.getType().equals(FormFieldTypes.DOUBLE) && formFieldValue instanceof String) {
+            else if (formField.getType().equals(FormFieldTypes.DECIMAL) && formFieldValue instanceof String) {
                 String strFieldValue = (String) formFieldValue;
                 if (StringUtils.isNotEmpty(strFieldValue) && NumberUtils.isNumber(strFieldValue)) {
                     result = Double.valueOf(strFieldValue);
